+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Digit rules can have multiple constraints.
 - Symbol rules can have multiple constraints.
 - Character class rules can have multiple constraints.
+- Have I Been Pwned? rules can have multiple constraints.
 
 ### Changed
 - Minimum constraint of password change rule is never null.